requirements.txt
fastapi
uvicorn
streamlit
python-multipart
langchain
langchain-community
ollama
sentence-transformers
chromadb
numpy
pypdf2
beautifulsoup4
requests
python-docx
faiss-cpu
python-dotenv
pydantic
aiofiles